Fixed batch dimension in MultivariateUniform prob computations.
Weights of samples are computed w.r.t the center of the node
independently of the sample location: f(x)q(c) / q(c)

This does not change the behavior of the Haar kernel tree but makes
the Gauss kernel tree work better.

Added self-normalization step after generating new samples, this self
norm step is important to preserve the volume. In the case of the
gaussian is specially sensitive if re-normalization is skipped.

Added visualization of the gaussian kernels.